requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
qrcode
Pillow
itsdangerous
python-multipart
sqlalchemy
pyotp
requests
psutil
bcrypt
pydantic[email]
PyJWT
polars
python-dotenv
dotenv